Fixes exception thrown for networkx recent version stamps
Parsing of the version string for Networkx relies on the assumption that there are only 2 parts (eg. 1.2 or 2.3). Releases which include subversions (2.5.1) fail to be parsed and raise an exception.

Since only the major part of the version number is needed, I have modified the method to split only this component. Additionally, I have removed the integer conversion, since there is no need for making the subsequent decesion on API usage.